pyproject.toml
[tool.poetry]
name = "taipy-demo-logistic-regression"
version = "0.1.0"
description = ""
authors = ["Grégoire Marabout-Demazure <gmarabout@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
taipy-core = "^2.0.4"
taipy-gui = "^2.0.3"
numpy = "^1.23.5"
scikit-learn = "^1.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
